requirements.txt
# File: requirements.txt

google-generativeai>=0.3.0
elevenlabs>=1.0.0,<2.0.0
manim>=0.17.3
python-dotenv>=1.0.0
numpy>=1.24.0,<2.3.0
Pillow>=9.5.0